+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.16.0] - 2025-08-29
+
 ### Breaking Changes
 
 - Drop support for legacy code and ReductStore API versions below 1.13
@@ -284,7 +286,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement Reduct Storage HTTP API v0.4, [PR-16](https://github.com/reduct-storage/reduct-py/pull/16)
 
-[Unreleased]: https://github.com/reduct-storage/reduct-py/compare/v1.15.0...HEAD
+[Unreleased]: https://github.com/reduct-storage/reduct-py/compare/v1.16.0...HEAD
+
+[1.16.0]: https://github.com/reduct-storage/reduct-py/compare/v1.15.0...v1.16.0
 
 [1.15.0]: https://github.com/reduct-storage/reduct-py/compare/v1.14.0...v1.15.0